# CS6650_Final_Project/final_mastery/album-store.openapi.yaml
openapi: "3.0.3"
info:
  title: ChaosArena v1 — Album Store
  version: "1.0"
  description: |
    Official API contract for CS 6650 ChaosArena v1-album-store.
    Implement every endpoint exactly as specified. Field names, types,
    and status strings are checked literally.
servers:
  - url: "http://your-service-url"
    description: Your deployed service
paths:
...
  /albums/{album_id}:
    parameters:
      - name: album_id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    put:
      summary: Create or update an album
      operationId: putAlbum
      description: Idempotent — calling twice with the same album_id must not create two records.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlbumRequest"
      responses:
        "200":
          description: Album updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
        "201":
          description: Album created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"

    get:
      summary: Get an album by ID
      operationId: getAlbum
      responses:
        "200":
          description: Album found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
        "404":
          description: Album not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Album:
      type: object
      required: [album_id, title, description, owner]
      properties:
        album_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        owner:
          type: string
      example:
        album_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        title: "My Summer Trip"
        description: "Photos from Cancun"
        owner: "student@northeastern.edu"
    AlbumRequest:
      type: object
      required: [album_id, title, description, owner]
      properties:
        album_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        owner:
          type: string
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
      example:
        error: "not found"