# CS6650_Final_Project/final_mastery/server/cache/redis.go
package cache

import (
	"context"
	"encoding/json"
	"errors"
	"fmt"
	"time"

	"github.com/redis/go-redis/v9"
)

type PhotoStatus struct {
	PhotoID string `json:"photo_id"`
	AlbumID string `json:"album_id"`
	Seq     int64  `json:"seq"`
	Status  string `json:"status"`
	URL     string `json:"url,omitempty"`
}

type Album struct {
	AlbumID     string `json:"album_id"`
	Title       string `json:"title"`
	Description string `json:"description"`
	Owner       string `json:"owner"`
}

type Cache struct {
	client *redis.Client
}

func New(redisURL string) (*Cache, error) {
	opt, err := redis.ParseURL(redisURL)
	if err != nil {
		return nil, fmt.Errorf("parse redis url: %w", err)
	}
	client := redis.NewClient(opt)
	if err := client.Ping(context.Background()).Err(); err != nil {
		return nil, fmt.Errorf("redis ping: %w", err)
	}
	return &Cache{client: client}, nil
}

// NextSeq atomically increments and returns the next seq for an album.
// Redis INCR is atomic — safe under any concurrency across all instances.
func (c *Cache) NextSeq(ctx context.Context, albumID string) (int64, error) {
	return c.client.Incr(ctx, "album:"+albumID+":seq").Result()
}

// InitSeq sets the seq counter for an album only if it is not already set (or below maxSeq).
// Used at startup to warm counters from DB max(seq).
func (c *Cache) InitSeq(ctx context.Context, albumID string, maxSeq int64) error {
	key := "album:" + albumID + ":seq"
	script := `
local cur = redis.call('GET', KEYS[1])
if not cur or tonumber(cur) < tonumber(ARGV[1]) then
  redis.call('SET', KEYS[1], ARGV[1])
end
return redis.call('GET', KEYS[1])`
	return c.client.Eval(ctx, script, []string{key}, maxSeq).Err()
}

// SetAlbum caches album metadata. No TTL — albums are immutable once written.
func (c *Cache) SetAlbum(ctx context.Context, a *Album) {
	b, err := json.Marshal(a)
	if err != nil {
		return
	}
	c.client.Set(ctx, "album:"+a.AlbumID, b, 0)
}

// GetAlbum retrieves cached album. Returns nil, nil on miss.
func (c *Cache) GetAlbum(ctx context.Context, albumID string) (*Album, error) {
	b, err := c.client.Get(ctx, "album:"+albumID).Bytes()
	if errors.Is(err, redis.Nil) {
		return nil, nil
	}
	if err != nil {
		return nil, err
	}
	var a Album
	if err := json.Unmarshal(b, &a); err != nil {
		return nil, err
	}
	return &a, nil
}

// SetPhotoStatus caches photo status with 1h TTL.
func (c *Cache) SetPhotoStatus(ctx context.Context, ps *PhotoStatus) {
	b, err := json.Marshal(ps)
	if err != nil {
		return
	}
	c.client.Set(ctx, "photo:"+ps.PhotoID, b, time.Hour)
}

// GetPhotoStatus retrieves cached photo status. Returns nil, nil on miss.
func (c *Cache) GetPhotoStatus(ctx context.Context, photoID string) (*PhotoStatus, error) {
	b, err := c.client.Get(ctx, "photo:"+photoID).Bytes()
	if errors.Is(err, redis.Nil) {
		return nil, nil
	}
	if err != nil {
		return nil, err
	}
	var ps PhotoStatus
	if err := json.Unmarshal(b, &ps); err != nil {
		return nil, err
	}
	return &ps, nil
}

// DeletePhotoStatus removes a photo from cache (called on DELETE).
func (c *Cache) DeletePhotoStatus(ctx context.Context, photoID string) {
	c.client.Del(ctx, "photo:"+photoID)
}